Compiled and check got the menu working.
I checked the main menu functiality and got it working to where we can use this to start controlling the relay board. I checked the wiring on the relay board and it should be good to go.

So far we have everything we need to really get some interesting stuff as far as pnumatics and motor testing is concerned.